+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- The ability to use `:silent` with the `:Compile` and `:Recompile` commands
+  - Also the ability to pass the `smods.silent` to the API functions
+
 ## [1.0.1] - 2023-10-27
 
 ### Fixed
@@ -21,4 +28,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The ability to use `:vert` with the `:Compile` and `:Recompile` commands
   - Also the ability to pass the `smods` field (and `smods.vertical`) to the API functions
 
+[unreleased]: https://github.com/ej-shafran/compile-mode.nvim/compare/v1.0.1...HEAD
 [1.0.1]: https://github.com/ej-shafran/compile-mode.nvim/compare/v1.0.0...v1.0.1